Refactor a few utility functions
